pyproject.toml
[project]
name = "analysisposts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "mcp[cli]>=1.22.0",
    "fastmcp>=2.13.0",
    "numpy>=2.3.5",
    "pandas>=2.2.0",
    "pocketflow>=0.0.3",
    "requests>=2.32.5",
    "zai-sdk>=0.0.4.2",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.2",
    "networkx>=3.3",
    "jieba>=0.42.1",
    "scikit-learn>=1.5.0",
    "streamlit>=1.31.0",
]

[project.scripts]
analysis = "main:main"

[tool.pytest.ini_options]
addopts = "--basetemp=.pytest_tmp"
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]
markers = ["integration: live LLM API connectivity"]

[dependency-groups]
dev = ["pytest>=8.0"]
